add test for login json
